+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+* Added extension on `List<DropDownItem>` (aka `DropDownList`) that allows for selection/deselection of items in the list. Also added `selected` and `unselected` getters to return subsets of the list depending on the selection status of the `DropDownItem`.
+* Added methods on `DropDownData` that match new extension methods and apply it to both its `data` and its `future` data.
+* Added `select()` and `deselect()` methods to `DropDownItem`.
+
 ## 0.0.4
 
 * Updated inline documentation.